Add listeners to run the bot and the game
Added a listener to run the bot, if this happens, the username and chat
ID will be recorded in the user base object. Added listener to start
the game, when this happens, there will be a check for another user
in the database, if he is, he will receive a message asking to start
the game. If there is no other user in the database, the corresponding
message will be sent to the first user.