Change the function of the loop, shot and game
Moved the callback query listener from the loop function to the game
function because the loop function is called more than once and so the
listeners multiply, which is wrong. Added game object fill in game
function. Added shots of successful and unsuccessful shots as a
function of the shot.